backend/requirements.txt
# Django EAS Backend Requirements
# Core Django
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3

# Database
psycopg2-binary==2.9.9
Pillow==10.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-rest-auth==0.9.5

# Development
python-decouple==3.8
django-extensions==3.2.3

# QR Code generation
qrcode[pil]==7.4.2

# Campus & Analytics
python-dateutil==2.8.2

# Testing
pytest-django==4.8.0
factory-boy==3.3.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0